OrderBookInrementBuilder.
